moved ValFormField from custom_types to widget. included an exception that will use base primitives for creating the database table. once database table is created will use correct val fields.
