+++ escalate/core/views/__init__.py
@@ -6,4 +6,4 @@ from .export_views import *
 from .model_tag import *
 from .workflow import *
 from .user_views import *
-from .misc_views import *
+#from .misc_views import *
